:zap: fit to mobile
