+++ src/components/Benefit/index.module.scss
@@ -1,18 +1,19 @@
 .benefit {
-  height: calc(100vh - var(--ifm-navbar-height));
+  // height: calc(100vh - var(--ifm-navbar-height));
   background: url("/img/page/bg.png") no-repeat;
   position: relative;
   background-size: 100% 100%;
   background-attachment: fixed;
+  margin-top: 80px;
 
   @media (max-width: 996px) {
     height: auto;
   }
 
   &__title {
-    font-family: 'Roboto';
+    font-family: 'kubegems';
     font-style: normal;
-    font-weight: 700;
+    font-weight: 400;
     font-size: 36px;
     line-height: 120%;
     /* identical to box height, or 58px */
@@ -79,10 +80,10 @@
     }
 
     &__title {
-      font-family: 'Roboto';
+      font-family: 'kubegems';
       font-style: normal;
-      font-weight: 600;
-      font-size: 18px;
+      font-weight: 400;
+      font-size: 24px;
       line-height: 150%;
       /* identical to box height, or 27px */
       text-align: center;
